refactor: Move test codes into trace package
