+++ advanced/src/test/java/hello/advanced/trace/logtrace/ThreadLocalLogTraceTest.java
@@ -1,12 +1,13 @@
-package hello.advanced.app.trace.logtrace;
+package hello.advanced.trace.logtrace;
 
 import hello.advanced.trace.TraceStatus;
 import hello.advanced.trace.logtrace.FieldLogTrace;
+import hello.advanced.trace.logtrace.ThreadLocalLogTrace;
 import org.junit.jupiter.api.Test;
 
-class FieldLogTraceTest {
+class ThreadLocalLogTraceTest {
 
-    FieldLogTrace trace = new FieldLogTrace();
+    ThreadLocalLogTrace trace = new ThreadLocalLogTrace();
 
     @Test
     void begin_end_level2() {
