+++ advanced/src/test/java/hello/advanced/trace/template/TemplateMethodTest.java
@@ -1,8 +1,8 @@
-package hello.advanced.app.trace.template;
+package hello.advanced.trace.template;
 
-import hello.advanced.app.trace.template.code.AbstractTemplate;
-import hello.advanced.app.trace.template.code.SubClassLogic1;
-import hello.advanced.app.trace.template.code.SubClassLogic2;
+import hello.advanced.trace.template.code.AbstractTemplate;
+import hello.advanced.trace.template.code.SubClassLogic1;
+import hello.advanced.trace.template.code.SubClassLogic2;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 
@@ -40,16 +40,16 @@ public class TemplateMethodTest {
      */
     @Test
     void templateMethodV1() {
-        AbstractTemplate template1 = new SubClassLogic1();
+        hello.advanced.trace.template.code.AbstractTemplate template1 = new SubClassLogic1();
         template1.execute();
 
-        AbstractTemplate template2 = new SubClassLogic2();
+        hello.advanced.trace.template.code.AbstractTemplate template2 = new SubClassLogic2();
         template2.execute();
     }
 
     @Test
     void templateMethodV2() {
-        AbstractTemplate template1 = new AbstractTemplate() {
+        hello.advanced.trace.template.code.AbstractTemplate template1 = new hello.advanced.trace.template.code.AbstractTemplate() {
             @Override
             protected void call() {
                 log.info("비즈니스 로직1 실행");
@@ -58,7 +58,7 @@ public class TemplateMethodTest {
         log.info("클래스 이름1={}", template1.getClass());
         template1.execute();
 
-        AbstractTemplate template2 = new AbstractTemplate() {
+        hello.advanced.trace.template.code.AbstractTemplate template2 = new AbstractTemplate() {
             @Override
             protected void call() {
                 log.info("비즈니스 로직2 실행");
