+++ advanced/src/test/java/hello/advanced/trace/template/code/SubClassLogic1.java
@@ -1,4 +1,4 @@
-package hello.advanced.app.trace.template.code;
+package hello.advanced.trace.template.code;
 
 import lombok.extern.slf4j.Slf4j;
