+++ advanced/src/test/java/hello/advanced/trace/threadlocal/code/FieldService.java
@@ -1,4 +1,4 @@
-package hello.advanced.app.trace.threadlocal.code;
+package hello.advanced.trace.threadlocal.code;
 
 import lombok.extern.slf4j.Slf4j;
